gitlab-data-service/requirements.txt
pyjwt
flakon
flask_webtest
cryptography
sqlalchemy
flask_sqlalchemy
chaussette
